Got my screen working (all but sprites)
This is hands-down the best version of my screen JS code that I've ver created; it's still very Commodore-like, but this time it sets up a canvas to fit whatever width/height I want.  The character set, screen RAM, and color RAM are all working, and technically this is enough graphics to build games.

But I still want to learn how C64 sprites worked, and how NES sprites worked, and see if either of their approaches would be doable.