Got the input system working (I think)
I went to www.keycode.info to test which keys map to which buttons on the iCade, and I think I've got it working (seems to be working in the browser, both with the iCade's actual controlls and with the arrow/number keys).

I think before I mess with sprites, I might try and write/test that assembly program! :D