Warp zone - spent some time in World 1-2 (lol)
* Re-added and re-tested my old sound library
* Re-added and re-tested my old text-to-speech library

I'm making an 'executive decision' (lol) for Level 1: No sprite.
This is a 1.0 release, and idk if anyone besides me will ever play with it.
The graphics system is already much more powerful than the system that inspired it.
Unlike the VIC-20, this emulated graphics chip supports:

* Redefining characters in RAM (no copying from ROM first)
* All 16 colors all the time (not 8 foreground and 8 multi-color background)
* Redefining what those 16 colors are (not even the C64 or DOS can do that)

And if the VIC-20 can do Donkey Kong, Ms. Pac-Man, Omega Race, Frogger '07 etc.,
then this system is clearly good enough for a 1.0 release of a hobby arcade OS. :D