Started on that first 'boss battle' lol
* Added my old timer and modem files (not really related, but will be useful down the road)
* Created a bunch of redefined characters

The copy redefined chars code is being a bit of a goober; probably because the chars I want to copy
are > 256 chars long and the registers won't go higher.  I think some extra mathing will be involved.
This is one of those times I'm sure grateful for higher-level languages where mathing is automated. :D