Found & added noise generators from a previous project
It's not connected to anything yet, but it works (lol).
I created it a few years ago based on this tutorial:

https://noisehack.com/generate-noise-web-audio-api/

Next I need to wire it up to some RAM; I'm thinking 4 noise generators,
where the 2 high bits are the type of noise (white, brown or pink),
and the lower 6 bits are the volume (0-50, each 1 being 2% volume to get 100).