Bug fixed, beat World 1-2!  On to storage...
The bug was not a bug at all: I had a check in there, if TTS volume is off, then don't bother trying to speak (saving bytes/cycles in actual RAM), lol.  I forgot all about that.  But I set up button 8 to speak, and it works (see 1-2.s and 1-2.iCade).

So next, I'll start working on my storage system:

* localStorage
* my old modem.js (worked in other 8-bit things I built)
* maybe experiment with auto-load from URL