Played the bonus round (and won) lol
Yes, I can code games for iCadeOS in C!  Of course I still don't know cc65's config language enough to use all the features cc65 has to offer, but it works.  This will come in super handy for stuff like floating-point numbers (which are not easily accomplished in Assembly), and with a little hacking I can get printf to work (using sprintf and then an Assembly print routine).  Anyway cc65 is officially an option now. :)